requirements-test.txt
pytest
pillow
numpy
python-dateutil
pytest-asyncio
aiohttp
